+++ nhood-engine-service-svc/src/main/resources/application.properties
@@ -1,1 +1,2 @@
-
+app.name=@project.artifactId@
+app.version=@project.version@
